First stage of reorg
