+++ dashboard/models.py
@@ -1,0 +1,9 @@
+from django.db import models
+
+# Create your models here.
+
+
+class Dashboard(models.Model):
+
+    greeting = models.CharField(max_length=100)
+    target = models.CharField(max_length=100)
